Move back test utilities to vdk-impala plugin
