+++ projects/vdk-plugins/vdk-impala/tests/functional/jobs/load_dimension_scd1_template_job/02_run_load_dimension_scd1_template.py
@@ -1,7 +1,7 @@
 # Copyright 2021-2023 VMware, Inc.
 # SPDX-License-Identifier: Apache-2.0
 from vdk.api.job_input import IJobInput
-from vdk.plugin.test_utils.util_funcs import setup_testing_check
+from functional import test_utility
 
 __author__ = "VMware, Inc."
 __copyright__ = (
@@ -12,7 +12,7 @@ __copyright__ = (
 def run(job_input: IJobInput) -> None:
     template_args = job_input.get_arguments()
     check = template_args.get("check")
-    check = setup_testing_check(check)
+    check = test_utility.setup_testing_check(check)
 
     job_input.execute_template(
         template_name="load/dimension/scd1",
