--- projects/vdk-plugins/vdk-impala/tests/functional/jobs/load_fact_snapshot_template_job/requirements.txt
@@ -1,1 +1,0 @@
-vdk-test-utils
